pyproject.toml
[project]
name = "rss-bridge"
description = "Minimalistic RSS content extractor."
requires-python = ">=3.12"
dependencies = [
  "fastapi[standard]",
  "readability-lxml",
  "requests",
  "IPython",
  "feedparser",
  "feedgen",
  "tqdm",
]
dynamic = ["version"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.envs.default.scripts]
dev = "fastapi dev {args:main.py --host 0.0.0.0 --port 8000}"

[tool.hatch.envs.test]
dependencies = [
    "pytest",
]